+++ package.json
@@ -7,7 +7,31 @@
     "lib/*"
   ],
   "type": "module",
-  "keywords": [],
+  "exports": {
+    ".": {
+      "default": "./lib/index.js"
+    },
+    "./decode": {
+      "default": "./lib/decode.js"
+    },
+    "./encode": {
+      "default": "./lib/encode.js"
+    },
+    "./decodeStream": {
+      "default": "./lib/decodeStream.js"
+    },
+    "./encodeStream": {
+      "default": "./lib/encodeStream.js"
+    }
+  },
+  "keywords": [
+    "utf8",
+    "wtf8",
+    "unicode",
+    "encode",
+    "decode",
+    "stream"
+  ],
   "author": "Joe Hildebrand <joe-github@cursive.net>",
   "license": "MIT",
   "repository": {
