+++ src/const.ts
@@ -1,0 +1,8 @@
+export const BOM = 0xfeff; // U+FEFF: ZERO WIDTH NO-BREAK SPACE
+export const EMPTY = new Uint8Array(0);
+export const MAX_HIGH_SURROGATE = 0xdbff;
+export const MAX_LOW_SURROGATE = 0xdfff;
+export const MIN_HIGH_SURROGATE = 0xd800;
+export const MIN_LOW_SURROGATE = 0xdc00;
+export const REPLACEMENT = 0xfffd; // U+FFFD: REPLACEMENT CHARACTER
+export const WTF8 = 'wtf-8';
